refactor(web): share provider quota view state
